requirements.txt
crewai
langchain-groq
python-dotenv
requests
crewai-tools